% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecwm.R
\name{ecwm}
\alias{ecwm}
\title{A logistic regression model for the ECWM accounting for One-Sayers}
\usage{
ecwm(prevalence, pyes, data, only1 = NULL, iter = 500)
}
\arguments{
\item{prevalence}{a \code{prevalence} in the form
\code{response ~ terms}, where \code{response} is the response
variable and \code{terms} describes the linear predictors. The
response variable should be numeric with the values 1 for
"One" and 2 for "Two".}

\item{pyes}{name of the variable with the individual probabilities
of answering "Yes" to the innocuous question.}

\item{data}{data frame with individual records containing the response
and \code{pyes} variables, and the predictors.}

\item{only1}{right hand formula for one-saying. The default is NULL means ECWM without one-saying. For the ECWM
with one-saying, should be in the form \code{~terms}.}

\item{iter}{the maximum number of iterations for the derivatives using the
method "Nelder-Mead". The Default is 500.}
}
\value{
A list with the following objects:
\describe{
\item{covsPrev}{data frame for the predictors of \code{prevalence}.}
\item{covsONLY}{data frame for the predictors of \code{only1}.}
\item{coefs}{data frame with the parameter estimates.}
\item{gof}{fit statistics including the log-likelihood, AIC and a likelihood-ratio
test against the intercept-only model.}
\item{fitted}{vector with the estimated probabilities.}
\item{designMat_Prev}{design matrix of \code{prevalence}.}
\item{designMat_ONLY}{design matrix of \code{only1}.}
\item{freqs}{data frame with the observed and estimated response frequencies.}
\item{hessian}{Hessian matrix of the parameter estimates.}
}
}
\description{
Fits a logistic regression to the
extended crosswise model that accounts for self-protective One-Sayers,
and predicts the probabilities of the sensitive characteristic and one-saying
conditional on the predictors.
}
